backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
google-generativeai
yt-dlp
sqlalchemy
aiofiles
python-multipart
httpx
